# Use Stream's BeginRead/WriteInternal from FileStream